requirements.txt
flask
anthropic
langchain
langchain_community
python-dotenv